requirements.txt
scancode-toolkit==31.0.1
beautifulsoup4==4.11.1
grimoirelab_toolkit==0.1.12
isort==5.10.1
jsonpath==0.82
perceval==0.17.17
perceval_weblate==0.1.2
PyMySQL==0.9.3
PyYAML==6.0
requests==2.26.0
scipy==1.7.1
setuptools==58.1.0
SQLAlchemy==1.3.24
tornado==6.1
tqdm==4.62.3
urllib3==1.25.8
python-rpm-spec==0.11
rarfile==4.0
GitPython==3.1.27
DBUtils==1.3